image preprocessing refactor, now is more fast
