+++ android/build.gradle
@@ -22,6 +22,10 @@ rootProject.allprojects {
         maven {
             url 'https://jitpack.io'
         }
+        maven{
+            name 'ossrh-snapshot'
+            url 'https://oss.sonatype.org/content/repositories/snapshots'
+        }
     }
 }
 
@@ -51,8 +55,12 @@ dependencies{
     //implementation (files('libs/tesseract4android-release.aar'))
     api(name:"tesseract4android-release", ext: "aar")
     implementation 'com.github.vladiH:opencv-android:v1.0.0'
-    implementation 'org.tensorflow:tensorflow-lite:2.7.0'
-    implementation 'org.tensorflow:tensorflow-lite-gpu:2.7.0'
-    implementation 'org.tensorflow:tensorflow-lite-support:0.3.1'
-    implementation 'org.tensorflow:tensorflow-lite-metadata:0.3.1'
+    implementation 'org.tensorflow:tensorflow-lite:2.10.0'
+    implementation 'org.tensorflow:tensorflow-lite-api:2.10.0'
+    implementation 'org.tensorflow:tensorflow-lite-gpu:2.10.0'
+    implementation 'org.tensorflow:tensorflow-lite-gpu-api:2.10.0'
+    implementation 'org.tensorflow:tensorflow-lite-gpu-delegate-plugin:0.4.3'
+    implementation 'org.tensorflow:tensorflow-lite-support:0.4.3'
+    implementation 'org.tensorflow:tensorflow-lite-metadata:0.4.3'
+    implementation 'org.tensorflow:tensorflow-lite-select-tf-ops:2.11.0'
 }
