+++ example/android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.vladih.computer_vision.flutter_vision_example">
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
    <application
+        android:requestLegacyExternalStorage="true"
         android:label="flutter_vision_example"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher"
